--- src/components/Detail/login/Login.jsx
@@ -1,8 +1,0 @@
-import React from 'react';
-import AccountForm from '../accountForm/AccountForm';
-
-function Login({ open, handleClose }) {
-    return <AccountForm isLogin open={open} handleClose={() => handleClose()} />;
-}
-
-export default Login;
